atulizando projeto

+++ src/main/java/br/com/estruturaDeDados/Aresta.java
@@ -1,3 +1,3 @@
 package br.com.estruturaDeDados;
 
-public record Aresta(Double distancia, Double tempo) {}
+public record Aresta(Double distancia, Integer tempo) {}
